added check if demo is already imported

+++ backend/api/routes/demo.js
@@ -2,6 +2,7 @@ const express = require('express')
 const router = express.Router()
 const parser = require('../parser/main')
 const { exec } = require('child_process')
+const fs = require('fs')
 
 /**
  * @swagger
@@ -29,37 +30,50 @@ router.get('/', (req, res, next) => {
 module.exports = router
 
 async function demoPresentation() {
-    return new Promise((resolve, reject) => {
-        return new Promise((resolve, reject) => {
-            // copy .zip to storage/all
-            exec(`cp ${process.env.FILE_PATH}/utils/Demo.zip ${process.env.FILE_PATH}/storage/all`, (err, stdout, stderr) => {
-                if (err) {
-                    console.log('Error on copy')
-                    reject({ status: 500, msg: 'Internal Server Error. Unable to copy to new directory' })
-                }
-                else {
-                    resolve()
-                }
+    return new Promise(async (resolve, reject) => {
+        var dir = `${process.env.FILE_PATH}/storage/5f18a16a4417a6565c3ecddf`
+        if(fs.existsSync(dir)){
+            await parser.executePresentation('5f18a16a4417a6565c3ecddf').then((response) =>{
+                resolve(response)
             })
-        }).then(async () => {
-            // import zip
-            await parser.importPresentation('Demo.zip')
-                .then(async (response) => {
-                    console.log('RESPONSE', response)
-                    // execute presentation
-                    await parser.executePresentation('5f18a16a4417a6565c3ecddf').then((response) =>{
-                        resolve(response)
+            .catch((err) =>{
+                reject(err)
+            })
+        }
+        else{
+            return new Promise((resolve, reject) => {
+                // copy .zip to storage/all
+                exec(`cp ${process.env.FILE_PATH}/utils/Demo.zip ${process.env.FILE_PATH}/storage/all`, (err, stdout, stderr) => {
+                    if (err) {
+                        console.log('Error on copy')
+                        reject({ status: 500, msg: 'Internal Server Error. Unable to copy to new directory' })
+                    }
+                    else {
+                        resolve()
+                    }
+                })
+            }).then(async () => {
+                // import zip
+                await parser.importPresentation('Demo.zip')
+                    .then(async (response) => {
+                        console.log('RESPONSE', response)
+                        // execute presentation
+                        await parser.executePresentation('5f18a16a4417a6565c3ecddf').then((response) =>{
+                            resolve(response)
+                        })
+                        .catch((err) =>{
+                            reject(err)
+                        })
                     })
-                    .catch((err) =>{
+                    .catch((error) => {
+                        console.log('EROOR', error)
                         reject(err)
                     })
-                })
-                .catch((error) => {
-                    console.log('EROOR', error)
-                    reject(err)
-                })
-        }).catch((err) =>{
-            reject(err)
-        })
+            }).catch((err) =>{
+                reject(err)
+            })
+        }
+
+        
     })
 }
